[Android] Add custom removal dialog for Home/Work profiles in settings
This change introduces a custom confirmation dialog text for removing
Home or Work profiles in Autofill settings. It also adds links within
the text that forward to the Google Account address editing site.
However, these links are currently not clickable because the dialog
needs to be migrated from AlertDialog to ModalDialog.

Mocks:
https://www.figma.com/design/IwB272Z5KxJyjfSwryZ0ms/Chrome-Autofill--Home---Work?node-id=3914-20785&m=dev

UI change:
http://screencast/cast/NDU1NjY3NTUzMTkzMTY0OHw5NGNlOTU4NC0xZQ

Bug: 436459634
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6819959
Commit-Queue: Piotr Kotynia <piotrkotynia@google.com>
Cr-Commit-Position: refs/heads/main@{#1496879}

+++ chrome/browser/autofill/android/java/src/org/chromium/chrome/browser/autofill/editors/EditorDialogView.java
@@ -110,7 +110,8 @@ public class EditorDialogView extends AlwaysDismissedDialog
     private @Nullable AlertDialog mConfirmationDialog;
 
     private @Nullable String mDeleteConfirmationTitle;
-    private @Nullable String mDeleteConfirmationText;
+    private @Nullable CharSequence mDeleteConfirmationText;
+    private @Nullable String mDeleteConfirmationPrimaryButtonText;
 
     private @Nullable Runnable mDeleteRunnable;
     private @Nullable Runnable mDoneRunnable;
@@ -181,8 +182,13 @@ public class EditorDialogView extends AlwaysDismissedDialog
         mDeleteConfirmationTitle = deleteConfirmationTitle;
     }
 
-    public void setDeleteConfirmationText(@Nullable String deleteConfirmationText) {
+    public void setDeleteConfirmationText(@Nullable CharSequence deleteConfirmationText) {
         mDeleteConfirmationText = deleteConfirmationText;
+    }
+
+    public void setDeleteConfirmationPrimaryButtonText(
+            @Nullable String deleteConfirmationPrimaryButtonText) {
+        mDeleteConfirmationPrimaryButtonText = deleteConfirmationPrimaryButtonText;
     }
 
     public void setAllowDelete(boolean allowDelete) {
@@ -260,9 +266,13 @@ public class EditorDialogView extends AlwaysDismissedDialog
         toolbar.setOnMenuItemClickListener(
                 item -> {
                     if (item.getItemId() == R.id.delete_menu_id) {
-                        if (mDeleteConfirmationTitle != null || mDeleteConfirmationText != null) {
+                        if (mDeleteConfirmationTitle != null
+                                && mDeleteConfirmationText != null
+                                && mDeleteConfirmationPrimaryButtonText != null) {
                             handleDeleteWithConfirmation(
-                                    mDeleteConfirmationTitle, mDeleteConfirmationText);
+                                    mDeleteConfirmationTitle,
+                                    mDeleteConfirmationText,
+                                    mDeleteConfirmationPrimaryButtonText);
                         } else {
                             handleDelete();
                         }
@@ -620,7 +630,7 @@ public class EditorDialogView extends AlwaysDismissedDialog
     }
 
     private void handleDeleteWithConfirmation(
-            @Nullable String confirmationTitle, @Nullable String confirmationText) {
+            String confirmationTitle, CharSequence confirmationText, String primaryButtonText) {
         LayoutInflater inflater = LayoutInflater.from(getContext());
         View body = inflater.inflate(R.layout.confirmation_dialog_view, null);
         TextView titleView = body.findViewById(R.id.confirmation_dialog_title);
@@ -628,6 +638,7 @@ public class EditorDialogView extends AlwaysDismissedDialog
         TextView messageView = body.findViewById(R.id.confirmation_dialog_message);
         messageView.setText(confirmationText);
 
+        // TODO(crbug.com/430218067): Migrate to modal dialog to make links clickable.
         mConfirmationDialog =
                 new AlertDialog.Builder(getContext(), R.style.ThemeOverlay_BrowserUI_AlertDialog)
                         .setView(body)
@@ -641,7 +652,7 @@ public class EditorDialogView extends AlwaysDismissedDialog
                                     }
                                 })
                         .setPositiveButton(
-                                R.string.delete,
+                                primaryButtonText,
                                 (dialog, which) -> {
                                     handleDelete();
                                     mConfirmationDialog = null;
